로그인기능 구현

+++ src/App.tsx
@@ -11,9 +11,10 @@ import ListsContainer from "./components/ListsContainer/ListsContainer";
 import { useTypedDispatch, useTypedSelector } from "./hooks/redux";
 import EditModal from "./components/EditModal/EditModal";
 import LoggerModal from "./components/LoggerModal/LoggerModal";
-import { deleteBoard } from "./store/slices/boardsSlice";
+import { deleteBoard, sort } from "./store/slices/boardsSlice";
 import { addLog } from "./store/slices/loggerSlice";
 import { v4 as uuidv4 } from "uuid";
+import { DragDropContext, DropResult } from "@hello-pangea/dnd";
 
 function App() {
   const dispatch = useTypedDispatch();
@@ -30,29 +31,71 @@ function App() {
 
   const handleDeleteBoard = () => {
     if (boards.length > 1) {
-      dispatch(deleteBoard({ boardId: activeBoardId }));
-      dispatch(
-        addLog({
-          logId: uuidv4(),
-          logMessage: `게시판 삭제하기: ${getActiveBoard.boardName}`,
-          logAuthor: "user",
-          logTimestamp: String(Date.now()),
-        })
+      const confirmDelete = window.confirm(
+        "정말로 이 게시판을 삭제하시겠습니까?"
       );
-      const newIndexToSet = () => {
-        const indexToBeDeleted = boards.findIndex(
-          (board) => board.boardId === activeBoardId
+      if (confirmDelete) {
+        dispatch(deleteBoard({ boardId: activeBoardId }));
+        dispatch(
+          addLog({
+            logId: uuidv4(),
+            logMessage: `게시판 삭제: ${getActiveBoard.boardName}`,
+            logAuthor: "user",
+            logTimestamp: String(Date.now()),
+          })
         );
+        const newIndexToSet = () => {
+          const indexToBeDeleted = boards.findIndex(
+            (board) => board.boardId === activeBoardId
+          );
 
-        return indexToBeDeleted === 0
-          ? indexToBeDeleted + 1
-          : indexToBeDeleted - 1;
-      };
+          return indexToBeDeleted === 0
+            ? indexToBeDeleted + 1
+            : indexToBeDeleted - 1;
+        };
 
-      setActiveBoardId(boards[newIndexToSet()].boardId);
+        setActiveBoardId(boards[newIndexToSet()].boardId);
+      }
     } else {
       alert("게시판이 하나 남아있어서 삭제할 수 없습니다.");
     }
+  };
+
+  const handleDragEnd = (result: DropResult) => {
+    const { destination, source, draggableId } = result;
+    if (!destination) return;
+
+    const sourceList = lists.filter(
+      (list) => list.listId === source.droppableId
+    )[0];
+
+    dispatch(
+      sort({
+        boardIndex: boards.findIndex(
+          (board) => board.boardId === activeBoardId
+        ),
+        droppableIdStart: source.droppableId,
+        droppableIdEnd: destination.droppableId,
+        droppableIndexStart: source.index,
+        droppableIndexEnd: destination.index,
+        draggableId,
+      })
+    );
+
+    dispatch(
+      addLog({
+        logId: uuidv4(),
+        logMessage: `태스크(${
+          sourceList.tasks.filter((task) => task.taskId === draggableId)[0]
+            .taskName
+        }) 이동: ${sourceList.listName} => ${
+          lists.filter((list) => list.listId === destination.droppableId)[0]
+            .listName
+        }`,
+        logAuthor: "user",
+        logTimestamp: String(Date.now()),
+      })
+    );
   };
 
   return (
@@ -64,7 +107,9 @@ function App() {
         setActiveBoardId={setActiveBoardId}
       />
       <div className={board}>
-        <ListsContainer lists={lists} boardId={getActiveBoard.boardId} />
+        <DragDropContext onDragEnd={handleDragEnd}>
+          <ListsContainer lists={lists} boardId={getActiveBoard.boardId} />
+        </DragDropContext>
       </div>
       <div className={buttons}>
         <button className={deleteBoardButton} onClick={handleDeleteBoard}>
